+++ src/stream_formats.rs
@@ -1,5 +1,6 @@
 use crate::CONFIG;
 use crate::api::Chapters;
+use crate::channel::VideoMetadata;
 use crate::{
     api::{Format, VideoInfo},
     list::{ListItem, SelectionList},
@@ -8,8 +9,7 @@ use std::fmt::Display;
 
 #[derive(Default)]
 pub struct Formats {
-    pub title: String,
-    pub id: String,
+    pub metadata: VideoMetadata,
     pub video_formats: SelectionList<Format>,
     pub audio_formats: SelectionList<Format>,
     pub formats: SelectionList<Format>,
@@ -20,10 +20,9 @@ pub struct Formats {
 }
 
 impl Formats {
-    pub fn new(title: String, id: String, video_info: VideoInfo) -> Self {
+    pub fn new(metadata: VideoMetadata, video_info: VideoInfo) -> Self {
         let mut formats = Self {
-            title,
-            id,
+            metadata,
             video_formats: SelectionList::new(video_info.video_formats),
             audio_formats: SelectionList::new(video_info.audio_formats),
             formats: SelectionList::new(video_info.format_streams),